requirements.txt
streamlit==1.36.0
requests>=2.32.0
python-dotenv>=1.0,<2.0
markdown>=3.6
fpdf2>=2.7
